+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release 1.15.0
 ## General
 - Fixed irrelevant rooms not being unloaded while spectating other players that led to higher network throughput
+- Replaced dropdown menu for Cape Colors with a textbox that accepts a hexcode of any possible color
+- Added a checkbox for Rainbow Cape that's only active during events
 ## Arena
 - Added new gamemode: Drown! Kill and survive to buy your escape. Cooperative or Competitive
 - Scoring update: Empty kills and friendly fire now subtract from the user (in Teams, the killer loses the points). Points are no longer granted to everyone else.